port/src/pdmain.c:203:2: error: initializer element is not a compile-time constant         gsSPDisplayList(&var800613a0),         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~ include/PR/gbi.h:1703:33: note: expanded from macro 'gsSPDisplayList' #define gsSPDisplayList(   dl)  gsDma1p(   G_DL,dl,0,G_DL_PUSH)                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ include/PR/gbi.h:1666:5: note: expanded from macro 'gsDma1p'     (unsigned int)(s)                                ^~~~~~~~~~~~~~~~~

+++ src/include/types.h
@@ -8,6 +8,10 @@
 #include "pads.h"
 #include "tiles.h"
 #include "gbi.h"
+
+#include "stdio.h"
+#include "strings.h"
+
 #ifndef PLATFORM_N64
 #include "platform.h"
 #endif
@@ -5357,7 +5361,7 @@ struct guncmd {
 	u8 type;
 	u8 unk01;
 	u16 unk02;
-	s32 unk04;
+	intptr_t unk04;
 };
 
 struct pakthing {
